Changed the signature of the `Rectangle`
	* Renamed `Test` folder  to `test`.

+++ src/com/tw/step8/assignment1/Rectangle.java
@@ -2,16 +2,16 @@ package com.tw.step8.assignment1;
 
 public class Rectangle {
 
-    private final int length;
-    private final int breadth;
+    private final double length;
+    private final double breadth;
 
-    public Rectangle(int length, int breadth) {
+    public Rectangle(double length, double breadth) {
 
         this.length = length;
         this.breadth = breadth;
     }
 
-    public int area() {
+    public double area() {
         return length * breadth;
     }
